お気に入り登録用の処理をViewModel側に移植。 コメントをいくつか追加。
